Adding configuration command to run glew makefile
Temporary solution. Depends on using bash to execute make command. That means if the environment is windows, this won't work (without windows 10 and linux subsystem for windows). Even if the environment has bash, the makefile contains further dependencies. The next solution is to use a docker image which contains all of these dependencies, thus the user simply needs cmake and docker installed after having checked out this repository.